+++ src/modules/admin/controllers/traits/UserTrait.php
@@ -2,16 +2,16 @@
 
 namespace davidhirtz\yii2\skeleton\modules\admin\controllers\traits;
 
-use davidhirtz\yii2\skeleton\modules\admin\models\forms\UserForm;
+use davidhirtz\yii2\skeleton\models\User;
 use Yii;
 use yii\web\ForbiddenHttpException;
 use yii\web\NotFoundHttpException;
 
 trait UserTrait
 {
-    protected function findUserForm(int $id, ?string $permissionName = null): UserForm
+    protected function findUser(int $id, ?string $permissionName = null): User
     {
-        if (!$user = UserForm::findOne((int)$id)) {
+        if (!$user = User::findOne($id)) {
             throw new NotFoundHttpException();
         }