+++ src/modules/admin/widgets/forms/traits/UserActiveFormTrait.php
@@ -2,25 +2,27 @@
 
 namespace davidhirtz\yii2\skeleton\modules\admin\widgets\forms\traits;
 
+use davidhirtz\yii2\skeleton\helpers\ArrayHelper;
 use davidhirtz\yii2\skeleton\helpers\Html;
 use davidhirtz\yii2\skeleton\models\Trail;
 use davidhirtz\yii2\skeleton\models\User;
 use davidhirtz\yii2\skeleton\modules\admin\models\forms\UserForm;
-use davidhirtz\yii2\skeleton\widgets\forms\DynamicRangeDropdown;
 use davidhirtz\yii2\skeleton\widgets\forms\TimezoneDropdown;
 use davidhirtz\yii2\timeago\Timeago;
 use Yii;
 use yii\widgets\ActiveField;
 
-trait UserFormTrait
+trait UserActiveFormTrait
 {
-    /**
-     * @uses User::getCountries()
-     */
-    public function countryField(array $options = ['options' => ['prompt' => '']]): ActiveField|string
+    public function countryField(array $options = []): ActiveField|string
     {
-        $options['inputOptions']['prompt'] ??= '';
-        return $this->field($this->model, 'country', $options)->widget(DynamicRangeDropdown::class);
+        if (!$this->model->isAttributeRequired('country')) {
+            $options['inputOptions']['prompt'] ??= '';
+        }
+
+        $items = $this->model->user::getCountries();
+
+        return $this->field($this->model, 'country', $options)->dropDownList($items);
     }
 
     public function emailField(array $options = []): ActiveField|string
@@ -28,12 +30,10 @@ trait UserFormTrait
         return $this->field($this->model, 'email', $options)->input('email');
     }
 
-    /**
-     * @uses User::getLanguages()
-     */
     public function languageField(array $options = []): ActiveField|string
     {
-        return $this->field($this->model, 'language', $options)->widget(DynamicRangeDropdown::class);
+        $items = ArrayHelper::getColumn($this->model->user::getLanguages(), 'name');
+        return $this->field($this->model, 'language', $options)->dropDownList($items);
     }
 
     public function newPasswordField(array $options = []): ActiveField|string
@@ -52,18 +52,6 @@ trait UserFormTrait
     }
 
     /**
-     * @uses User::getStatuses()
-     */
-    public function statusField(array $options = []): ActiveField|string
-    {
-        if ($this->model->isOwner()) {
-            return '';
-        }
-
-        return $this->field($this->model, 'status', $options)->widget(DynamicRangeDropdown::class);
-    }
-
-    /**
      * @uses User::getTimezones()
      */
     public function timezoneField(array $options = []): ActiveField|string
@@ -79,11 +67,11 @@ trait UserFormTrait
 
     protected function getPicturePreview(): string
     {
-        if (!$this->model->picture) {
+        if (!$this->model->user->picture) {
             return '';
         }
 
-        return $this->row($this->offset(Html::img($this->model->getPictureUrl(), [
+        return $this->row($this->offset(Html::img($this->model->user->getPictureUrl(), [
             'style' => 'max-width:150px',
         ])));
     }
@@ -97,26 +85,34 @@ trait UserFormTrait
 
     protected function getFooterItems(): array
     {
+        $user = $this->model->user;
         $items = [];
 
-        if (!$this->model->getIsNewRecord()) {
-            if ($this->model->updated_at) {
+        if (!$user->getIsNewRecord()) {
+            if ($user->updated_at) {
                 $hasTrailAuth = Yii::$app->getUser()->can('trailIndex');
+
                 $text = Yii::t('skeleton', 'Last updated {timestamp}', [
-                    'timestamp' => Timeago::tag($this->model->updated_at),
+                    'timestamp' => Timeago::tag($user->updated_at),
                 ]);
 
-                $items[] = $hasTrailAuth ? Html::a($text, Trail::getAdminRouteByModel(User::instance(), $this->model->id)) : $text;
+                $items[] = $hasTrailAuth
+                    ? Html::a($text, Trail::getAdminRouteByModel(User::instance(), $user->id))
+                    : $text;
             }
 
-            if ($this->model->created_by_user_id) {
+            if ($user->created_by_user_id) {
+                $route = Yii::$app->getUser()->can(User::AUTH_USER_UPDATE, ['user' => $user->admin])
+                    ? ['/admin/user/update', 'id' => $user->id]
+                    : null;
+
                 $items[] = Yii::t('skeleton', 'Created by {user} {timestamp}', [
-                    'timestamp' => Timeago::tag($this->model->created_at),
-                    'user' => Html::username($this->model->admin, Yii::$app->getUser()->can(User::AUTH_USER_UPDATE, ['user' => $this->model->admin]) ? ['/admin/user/update', 'id' => $this->model->id] : null),
+                    'timestamp' => Timeago::tag($user->created_at),
+                    'user' => Html::username($user->admin, $route),
                 ]);
             } else {
                 $items[] = Yii::t('skeleton', 'Signed up {timestamp}', [
-                    'timestamp' => Timeago::tag($this->model->created_at),
+                    'timestamp' => Timeago::tag($user->created_at),
                 ]);
             }
         }